rm v1/2, add Dockerfile
