+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src-new/client/main.tsx"></script>
+    <script type="module" src="/src/client/main.tsx"></script>
   </body>
 </html>
