+++ package.json
@@ -1,11 +1,15 @@
 {
   "name": "cycling-geometry-analyzer",
   "private": true,
-  "version": "0.0.0",
+  "version": "3.0.0",
   "type": "module",
+  "engines": {
+    "node": ">=20.0.0"
+  },
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
+    "start": "vite preview --host 0.0.0.0 --port 80",
     "preview": "vite preview"
   },
   "dependencies": {
