--- src-new/client/app.css
@@ -1,13 +1,0 @@
-.cartesian-svg line.grid-line {
-  stroke: #ddd;
-  stroke-width: 1;
-}
-
-.cartesian-svg line.grid-center-line {
-  stroke: #000;
-  stroke-width: 2;
-}
-
-.debug {
-  display: none
-}
