+++ src/client/app.css
@@ -1,189 +1,13 @@
-#app {
-  max-width: 1000px;
-  margin: 0 auto;
-  width:100%;
-  text-align: center;
-  border: thin solid black;
-
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.preact:hover {
-  filter: drop-shadow(0 0 2em #673ab8aa);
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
-
-.workspace-wrapper {
-  max-width: 1200px;
-  width: 100%;
-  height: 500px;
-  display: block;
-  margin: 0 auto;
-  position: relative;
-}
-
-svg.workspace {
-  background: #f4f4f4;
-  border: 1px solid red;
-  display: block;
-  margin: auto;
-  height: 100%;
-  width: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 0;
-}
-
-svg.workspace .axis-line {
-  stroke: lightgrey;
+.cartesian-svg line.grid-line {
+  stroke: #ddd;
   stroke-width: 1;
 }
 
-svg.workspace .grid-line {
-  stroke: #e6e3e3;
-  stroke-width: 1;
-}
-
-.stem {
-  display: flex;
-  flex-direction: row;
-  justify-content: start;
-  position: absolute;
-  top: 0;
-  left:0;
-  right: 0;
-  bottom: 0;
-  margin: auto;
-  width: 100%;
-}
-
-svg.stem-coords {
-  flex-grow: 1;
-  overflow: visible;
-}
-
-.cartesian-svg #stem-shaft-drawing {
-  fill: none;
-  z-index: 100;
-}
-
-.cartesian-svg line {
-  stroke: purple;
+.cartesian-svg line.grid-center-line {
+  stroke: #000;
   stroke-width: 2;
 }
 
-.stem-half {
-  /* border: 1px solid orange; */
-  height: auto;
-  top: 0;
-  width: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.stem-half.stem-back {
-  left: 0;
-}
-
-.stem-half.stem-front {
-  right: 0;
-}
-
-.stem-half img {
-  width: 100%;
-  height: auto;
-}
-
-.stem-half .stem-plot {
-  position: relative;
-}
-
-.stem-half.stem-back .plot-point {
-  right: 0;
-}
-
-.stem-half.stem-front .plot-point {
-  left: 0;
-}
-
-.stem-half .plot-point, .stem-half .center-point {
-  display: block;
-  /* display: none; */
-  padding: 3px;
-  width: 0;
-  height: 0;
-  position: absolute;
-  border-radius: 100%;
-  opacity: 0;
-}
-
-.stem-half .vertical-center {
-  position: absolute;
-  height: 100%;
-  width: 1px;
-  background: red;
-  top: 0;
-  padding: 0;
-}
-
-.stem-back .vertical-center {
-  left: 42%;
-}
-
-.stem-front .vertical-center {
-  left: 57%;
-}
-
-.center-point {
-  background: purple;
-  margin: auto 0;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-}
-
-.top-point {
-  background: red;
-  top: 0;
-}
-
-.bottom-point {
-  background: blue;
-  bottom: 0;
-}
-
-.stem-back .top-point {
-  top: 2%;
-}
-
-.stem-back .bottom-point {
-  bottom: 7%;
-}
-
-.stem-front .top-point {
-  top: 14%;
-}
-
-.stem-front .bottom-point {
-  bottom: 18%;
-}
-
-polygon {
-  stroke: black;
-  stroke-width: 1px;
+.debug {
+  display: none
 }
