+++ src/client/components/Stem/MergedStemFragments.tsx
