+++ src/client/components/Stem/Stem.tsx
