+++ src/client/components/Stem/StemFragment.tsx
