+++ src/client/configs/thomson.tsx
