+++ src/client/main.tsx
@@ -2,9 +2,8 @@ import { render } from 'preact'
 import App from './app.tsx'
 import './index.css'
 
-import stem from './configs/stem.tsx'
 
 (async () => {
   const root = document.getElementById('app');
-  render(<App stemData={stem} />, root)
+  render(<App />, root)
 })();
