+++ src/client/utils/calculations.tsx
