+++ src/client/utils/drawings.tsx
